apps/trade-backend/pyproject.toml
[tool.poetry]
name = "trade-backend"
version = "0.1.0"
description = ""
authors = ["Junho"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.109.0"
uvicorn = {extras = ["standard"], version = "0.27.0"}
sqlalchemy = "2.0.25"
asyncpg = "0.29.0"
pydantic-settings = "2.1.0"
websockets = "12.0"
kafka-python-ng = "2.2.2"
httpx = "0.26.0"
cachetools = "5.3.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
psycopg2-binary = "^2.9.9"
pandas = "^2.2.0"
numpy = "^1.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
